chore: update conda env and test dummy models (#57)
Because

our API produces inference output in standardized format when a model is labeled with a specific task such as classification (`TASK_CLASSIFICATION`) and detection (`TASK_DETECTION`). In order to make this happen, a user has to follow design guidelines to create the post-processing python model **from scratch**, so the model output can be converted to the standardized format. 

To streamline the post-processing python model creation process, a user can define the model using the [`triton-python-model` package in the triton-conda-env image](https://github.com/instill-ai/triton-python-model).

This commit

- update `triton-conda-env` to `v0.2.1-alpha`
- update the test dummy models to use `triton-python-model` package to define the post-processing model
- add [`yolov4-onnx-cpu-v2.zip`](https://artifacts.instill.tech/vdp/sample-models/yolov4-onnx-cpu-v2.zip), an updated version of `yolov4-onnx-cpu.zip` that uses `triton-python-model` package for the post-processing model