feat(model): Support New Fields for Multi-Modal Model In Text Generation Task and Refactor Existing Ones (#448)
**Because**

To Support Multi-Modal Like Llava in Text Generation Task  

**This pull request** introduces several enhancements to the text
generation task, improving its flexibility and usability. The key
changes made include:

1. **Refactoring of `stop_words` Field:**
- The `stop_words` field has been refactored to improve its
functionality and clarity.

2. **Introduction of `extra_params` Field:**
- A new `extra_params` field has been introduced. This field allows for
the inclusion of additional input parameters for the `generate` method,
providing greater control over the text generation process.

3. **Introduction of `prompt_image` Field:**
- To support multimodal text generation with image input, the
`prompt_image` field has been added. This allows for the inclusion of
images in the text generation process.

4. **Introduction of `temperature` Field:**
- The `temperature` field has been added to control the randomness of
text generation. This parameter allows users to adjust the temperature
of the generation process.

5. **Field Renaming:**
   - Several field names have been renamed for consistency and clarity:
     - `query` → `prompt`
     - `len` → `max_new_tokens`
     - `runtome_top_k` → `top_k`

6. **Removal of `bad_words` Field:**
- The `bad_words` field has been removed as it is not supported by every
model and may not be applicable in all scenarios.

These changes collectively enhance the text generation task, making it
more versatile and user-friendly.

The changes mentioned above depend on the changes made in this
[PR](https://github.com/instill-ai/protobufs/pull/221), which introduces
support for new fields for multimodal text generation in the protobuf
project.

---------

Co-authored-by: Heiru Wu <heiru.wu@instill.tech>